+++ libraries/wuss/window/create.c
@@ -109,6 +109,9 @@ result_t wuss_window_create(wuss_task_t        *task,
 
   win->task = task;
 
+  bg.colour     = wuss__resolve_colour(wuss, bg.colour);
+  bg.pattern_bg = wuss__resolve_colour(wuss, bg.pattern_bg);
+
   if (wuss__validate_backdrop(wuss, &bg) != result_OK)
   {
     wuss__free(wuss, win);