+++ libraries/wuss/window/set-background.c
@@ -7,6 +7,9 @@ result_t wuss_window_set_background(wuss_window_t  *window,
 {
   box_t content;
 
+  bg.colour     = wuss__resolve_colour(window->wuss, bg.colour);
+  bg.pattern_bg = wuss__resolve_colour(window->wuss, bg.pattern_bg);
+
   if (wuss__validate_backdrop(window->wuss, &bg) != result_OK)
     return result_WUSS_BAD_COLOUR;